refactor: 지하철 관련 오버레이 분리
